создал абстрактный компонент
